refactor(random): fix package name and some other errors
