+++ passgen/random/src/main/kotlin/dev/msfjarvis/aps/passgen/random/RandomPhonemesGenerator.kt
@@ -2,18 +2,12 @@
  * Copyright © 2014-2021 The Android Password Store Authors. All Rights Reserved.
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package dev.msfjarvis.aps.util.pwgen
+package dev.msfjarvis.aps.passgen.random
 
-import dev.msfjarvis.aps.passgen.random.`PasswordGenerator.kt`
-import dev.msfjarvis.aps.passgen.random.secureRandomBiasedBoolean
-import dev.msfjarvis.aps.passgen.random.secureRandomBoolean
-import dev.msfjarvis.aps.passgen.random.secureRandomCharacter
-import dev.msfjarvis.aps.passgen.random.secureRandomElement
-import dev.msfjarvis.aps.passgen.random.secureRandomNumber
-import dev.msfjarvis.aps.util.extensions.hasFlag
+import dev.msfjarvis.aps.passgen.random.util.hasFlag
 import java.util.Locale
 
-object RandomPhonemesGenerator {
+internal object RandomPhonemesGenerator {
 
   private const val CONSONANT = 0x0001
   private const val VOWEL = 0x0002
@@ -69,7 +63,7 @@ object RandomPhonemesGenerator {
     val upperCase = str.uppercase(Locale.ROOT)
     val lowerCase = str.lowercase(Locale.ROOT)
     val length = str.length
-    val isAmbiguous = str.any { it in `PasswordGenerator.kt`.AMBIGUOUS_STR }
+    val isAmbiguous = str.any { it in PasswordGenerator.AMBIGUOUS_STR }
   }
 
   /**
