+++ passgen/random/src/main/kotlin/dev/msfjarvis/aps/passgen/random/util/Extensions.kt
@@ -1,5 +1,7 @@
+package dev.msfjarvis.aps.passgen.random.util
+
 /** Clears the given [flag] from the value of this [Int] */
-internal fun Int.clearFlag(flag: Int): Int {
+internal infix fun Int.clearFlag(flag: Int): Int {
   return this and flag.inv()
 }
